requirements.txt
freetype-py==1.1
